remove stupid implementation of boot_timer.
