+++ debian/educloud.conf
@@ -9,9 +9,6 @@ mem = 0
 max_pboot_vms = 40
 max_pboot_delay = 30
 
-boot_timer = 45
-enable_boot_timer = 1
-
 [server]
 cpu_usage = 10
 cpu = 2
